python auto detects image pixel size

+++ embedded/src/image_filter/image_filter.c
@@ -1,8 +1,8 @@
 #include "image_filter.h"
 #include <assert.h>
 
-const int16_t sobel_kernel_x[3 * 3] = {-1, 0, 1, -2, 0, 2, -1, 0, 1};
-const int16_t sobel_kernel_y[3 * 3] = {-1, -2, -1, 0, 0, 0, 1, 2, 1};
+const int8_t sobel_kernel_x[3 * 3] = {-1, 0, 1, -2, 0, 2, -1, 0, 1};
+const int8_t sobel_kernel_y[3 * 3] = {-1, -2, -1, 0, 0, 0, 1, 2, 1};
 
 void imf_fill(ImageMatrix mat, IMF_TYPE value) {
     FOR_EACH_ELEMENT(mat) { ELEMENT(mat, row, col) = value; }
